pyproject.toml
[project]
name = "transaction_service"
version = "0.1.0"
description = "Blockchain transaction microservice for TraderJoe swaps on Avalanche"
authors = [
    { name = "Mübarek", email = "0xmillennium@protonmail.com" }
]
readme = "README.md"
requires-python = ">=3.10"

# `pip install .` komutu çalıştığında kurulacak ana bağımlılıklar
dependencies = [
    "pydantic[email] ~=2.10.0",
    "fastapi ~=0.115.0",
    "sqlalchemy ~=2.0.36",
    "aio-pika ~=9.5.5",
    "starlette ~=0.41.0",
    "aiormq ~=6.8.1",
    "PyYAML ~=6.0.2",
    "uvicorn ~=0.34.2",
    "argon2-cffi ~=23.1.0",
    "psycopg2 ~=2.9.10",
    "asyncpg ~=0.30.0",
    "web3 ~=7.6.0",
    "eth-account ~=0.13.1",
    "cryptography ~=44.0.0",
    "websockets ~=13.1",
    "python-jose[cryptography] ~=3.3.0",
    "httpx[http2] ~=0.28.0"
]

[project.optional-dependencies]
dev = [
    "pytest ~=8.3.5"
]

[project.urls]
"Homepage" = "soon..."
"Repository" = "soon..."

[tool.setuptools]
# setuptool ile kurulum yaparken kaynak kodu buradan al
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
